RR-173: Refactor form fillers seperating event form Refactor assertions to simplify form assertion
